+++ include/core/sha2.h
@@ -56,15 +56,15 @@ extern "C" {
 #endif
 
 typedef struct {
-    unsigned int tot_len;
-    unsigned int len;
+	size_t tot_len;
+	size_t len;
     unsigned char block[2 * SHA256_BLOCK_SIZE];
     uint32 h[8];
 } sha256_ctx;
 
 typedef struct {
-    unsigned int tot_len;
-    unsigned int len;
+	size_t tot_len;
+	size_t len;
     unsigned char block[2 * SHA512_BLOCK_SIZE];
     uint64 h[8];
 } sha512_ctx;
@@ -74,30 +74,30 @@ typedef sha256_ctx sha224_ctx;
 
 void sha224_init(sha224_ctx *ctx);
 void sha224_update(sha224_ctx *ctx, const unsigned char *message,
-                   unsigned int len);
+	size_t len);
 void sha224_final(sha224_ctx *ctx, unsigned char *digest);
-void sha224(const unsigned char *message, unsigned int len,
+void sha224(const unsigned char *message, size_t len,
             unsigned char *digest);
 
 void sha256_init(sha256_ctx * ctx);
 void sha256_update(sha256_ctx *ctx, const unsigned char *message,
-                   unsigned int len);
+                   size_t len);
 void sha256_final(sha256_ctx *ctx, unsigned char *digest);
-void sha256(const unsigned char *message, unsigned int len,
+void sha256(const unsigned char *message, size_t len,
             unsigned char *digest);
 
 void sha384_init(sha384_ctx *ctx);
 void sha384_update(sha384_ctx *ctx, const unsigned char *message,
-                   unsigned int len);
+	size_t len);
 void sha384_final(sha384_ctx *ctx, unsigned char *digest);
-void sha384(const unsigned char *message, unsigned int len,
+void sha384(const unsigned char *message, size_t len,
             unsigned char *digest);
 
 void sha512_init(sha512_ctx *ctx);
 void sha512_update(sha512_ctx *ctx, const unsigned char *message,
-                   unsigned int len);
+	size_t len);
 void sha512_final(sha512_ctx *ctx, unsigned char *digest);
-void sha512(const unsigned char *message, unsigned int len,
+void sha512(const unsigned char *message, size_t len,
             unsigned char *digest);
 
 #ifdef __cplusplus
